@renerocksai 2024-09-20
BTW: I am aware that taking perf measures on a mac is not what I usually do. I just don't have access to that Linux box ATM.