@zigster64 2024-09-20
Super interesting!

im getting similar httpz numbers on both M2 pro and Ryzen 5/linux

but nowhere near 300 for zap - its more ballpark with the others. Impressive 

you are making me want to get an M3 max !