@karlseguin 2024-09-21
It is weird how different my results are.

On an m2:

```
./wrk/measure.sh httpz
========================================================================
                          httpz
========================================================================
Running 10s test @ http://127.0.0.1:3000
  4 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.10ms    5.27ms 106.64ms   96.99%
    Req/Sec    61.13k    28.18k  241.88k    83.84%
  Latency Distribution
     50%    1.56ms
     75%    1.85ms
     90%    2.01ms
     99%   25.35ms
  2426670 requests in 10.10s, 124.97MB read
  Socket errors: connect 0, read 386, write 0, timeout 0
Requests/sec: 240254.03
Transfer/sec:     12.37MB
```

```
./wrk/measure.sh zig-zap 
========================================================================
                          zig-zap
========================================================================
Running 10s test @ http://127.0.0.1:3000
  4 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.57ms    8.69ms 160.88ms   95.37%
    Req/Sec    58.60k    35.92k  250.13k    74.94%
  Latency Distribution
     50%  665.00us
     75%    1.13ms
     90%    3.79ms
     99%   45.87ms
  2327928 requests in 10.09s, 352.99MB read
  Socket errors: connect 0, read 387, write 0, timeout 0
Requests/sec: 230727.51
Transfer/sec:     34.99MB
```


On an E3-1275 v6

```
========================================================================
                          httpz
========================================================================
Running 10s test @ http://127.0.0.1:3000
  4 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.28ms  377.20us  21.02ms   98.00%
    Req/Sec    78.78k     3.40k  103.86k    74.00%
  Latency Distribution
     50%    1.26ms
     75%    1.29ms
     90%    1.37ms
     99%    1.53ms
  3136631 requests in 10.03s, 161.53MB read
Requests/sec: 312587.49
```


```
========================================================================
                          zig-zap
========================================================================
Running 10s test @ http://127.0.0.1:3000
  4 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.34ms    1.92ms  54.45ms   95.66%
    Req/Sec    93.75k     3.88k  111.99k    91.00%
  Latency Distribution
     50%    1.00ms
     75%    1.09ms
     90%    1.41ms
     99%   10.63ms
  3732001 requests in 10.05s, 565.90MB read
Requests/sec: 371458.89
Transfer/sec:     56.33MB
```


For a "hello world" example, the main thing you can do is tweak the worker count and thread pool size. But if you're running `wrk` on the same machine as the server, I don't think you have any cores to spare. I tried various settings for both, and they largely just hurt performance:

```zig
    var server = try httpz.Server(void).init(allocator, .{
        .port = 3000,
        .workers = .{.count = 4},
        .thread_pool = .{.count = 4},
    }, {});
```