@renerocksai 2024-09-22
> It is weird how different my results are.
> 
> On an m2:

[...]

Awesome! Thanks for trying it with your configurations! The httpz Linux Transfer/sec readings would have been interesting, got cut off, but nvm.

Looking at the differences on a Linux machine, httpz and zap don't seem that far off. Those are the only numbers that really matter IMHO because if you are serious about a server, you don't run it on a Mac - might be a hot take IDK.

> 
> For a "hello world" example, the main thing you can do is tweak the worker count and thread pool size. But if you're running `wrk` on the same machine as the server, I don't think you have any cores to spare. I tried various settings for both, and they largely just hurt performance:
> 
> ```zig
>     var server = try httpz.Server(void).init(allocator, .{
>         .port = 3000,
>         .workers = .{.count = 2},
>         .thread_pool = .{.count = 6},
>     }, {});
> ```

Yeah, you have to be careful not to allocate more cores than you have - and cores are not created equal, esp. on new Macs.